Fix bug related to badger version
